src/requirements.txt
Jinja2==3.0.3
MarkupSafe==2.0.1
Pillow==8.4.0
PyYAML==5.4
aiohttp==3.8.1
async-timeout==4.0.1
attrs==21.2.0
blis==0.7.5
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
configparser==3.5.0
conllu==4.4.1
cycler==0.11.0
cymem==2.0.6
elasticsearch==7.15.2
filelock==3.4.0
fonttools==4.28.2
frozendict==2.1.0
frozenlist==1.2.0
fsspec==2021.11.1
gensim==4.1.2
h5py==3.6.0
huggingface-hub==0.0.19
idna==3.3
joblib==1.1.0
jsonpickle==2.0.0
kiwisolver==1.3.2
lxml==4.6.4
matplotlib==3.5.2
msgpack-numpy==0.4.7.1
msgpack==1.0.2
multidict==5.2.0
multiprocess==0.70.12.2
murmurhash==1.0.6
nltk==3.6.5
numpy==1.21.4
packaging==21.3
pandas==1.3.4
parse==1.19.0
pathy==0.6.1
patool==1.12
plotly==5.6.0
preshed==3.0.6
protobuf==3.19.4
psutil==5.8.0
pyarrow==6.0.1
pydantic==1.8.2
pyparsing==3.0.6
python-Levenshtein==0.12.2
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.2
scipy==1.7.0
scispacy==0.4.0
sentencepiece==0.1.96
setuptools==59.2.0
six==1.16.0
smart-open==5.2.1
spacy-legacy==3.0.8
spacy==3.0.7
srsly==2.4.2
tabulate==0.8.3
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.1
torch==1.9.1
torchvision==0.10.1
tqdm==4.48.2
transformers==4.11.3
typer==0.3.2
typing-extensions==4.0.0
urllib3==1.26.7
wasabi==0.8.2
xxhash==2.0.2
zensols.deeplearn==1.2.0
zensols.deepnlp==1.1.2
zensols.install==0.0.8
zensols.mimic==0.0.1
zensols.nlp==1.1.1
zensols.util==1.9.0
